requirements.txt
--find-links  https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
facenet_pytorch==2.5.2
numpy==1.18.0
pandas==1.3.5
torchvision==0.10.0+cpu
pillow==7.2.0
requests==2.23.0
wget==3.2
starlette

fastapi==0.54.1
uvicorn==0.11.1
gunicorn==20.0.4


